filter-repo: add a flag for restricting where --replace-text applies
Add a new callback, --replace-text-filename-callback, which will return
a simple boolean for any given filename.  When this command line flag is
passed, only filenames for which the callback returns true will have the
--replace-text rules applied to them.

I could also potentially simplify bfg-ish quite a bit by making use of
this new callback, but it serves as a pretty interesting demonstration
of how to do more complex operations so I have left it as-is for now.

NOTE: This should be modified to check for any --source parameter and
make sure the cat-file --batch process runs with the --source directory
as its current-working-directory.

Signed-off-by: Elijah Newren <newren@gmail.com>