filter-repo: make fresh clone warning scarier
Apparently, despite the fact that *overwrite* *repo* *history* are three
important words that each individually convey a lot of important
meaning, people ignore it and instinctively add --force.  Insert the
word "destructively" to get people to pause.

Further, change the end of the warning not to how to get around the
warning with the current repository, but instead with a suggestion that
they should instead be operating on a fresh clone and only then make a
side comment that the --force flag can be used to override.

Signed-off-by: Elijah Newren <newren@gmail.com>