filter-repo: make `--paths` as an option throw an error
People keep mistyping the `--path` argument as `--paths`.
Unfortunately, since python's argparse tries to helpfully autocomplete
unique abbreviations of arguments, this turns into `--paths-from-file`,
which is not what users wanted or expected.  And they don't realize the
problem but instead file bug reports saying that filter-repo is just
nuking all their files -- because the argument to `--paths` does exist
but its contents doesn't name any files in the users' repository.

To avoid this autocompletion, make `--paths` be an actual explicit
argument that argparse recognizes, and then in parse_args() just throw
an error if the user uses it.

Signed-off-by: Elijah Newren <newren@gmail.com>