filter-repo: fix misleading help message for --replace-message
When --replace-message was implemented in 0e63a81f344e (filter-repo: add
new --replace-message option, 2021-08-23), the documentation for the
option in git-filter-repo.txt was accurate, but the help message given
to argparse was missing two important words causing it to be misleading.
Since this option affects both commit AND tag messages, the help message
should mention both instead of implying that only commit messages are
affected.  Fix the oversight.

Signed-off-by: Elijah Newren <newren@gmail.com>