filter-repo: Do not require --force when using git-replace
Although using git-replace provided an almost-really-nice UI for
allowing simple parent rewriting to occur, the fact that `git replace
--graft` would result in creating a loose object which would then
trigger the not-a-fresh-clone error and cause users to need to specify
`--force` to override always really bothered me.  I don't like using
`--force` when it's not necessary, as it results in overuse and then
people risking losing data on something that isn't a fresh clone.

So, let's loosen the fresh-clone check to allow loose objects together
with a pack, but only if all loose objects happen to be objects that
replace refs point to.

Further, tweak the existing test cases using `git replace` to verify
that we only need to use the `--proceed` flag, and update the
documentation to point out that `--force` is no longer necessary.

Signed-off-by: Elijah Newren <newren@gmail.com>