filter-repo: change the --replace-refs default to 'delete-no-add'
I know, I just recently changed the default to 'update-no-add', but
upon reflection, that default was written assuming that replace refs
were mostly used by users to replace commits that didn't exist with
commits that did, i.e. as a mapping from a previous set of commits to
new current ones.  That's kind of rare; filter-repo might have been
the only thing to ever use replace refs that way, and since I'm making
the creation of such refs no longer the default, I don't think we'll see
it in the wild much.  Much more likely is folks using it to replace
existing commits with others, in which case we just want to make the
rewrite permanent and nuke the replace ref.  In fact, we have a few
examples in the manual that do precisely this.

So, I think my default change in eab90e9c1075 (filter-repo: change the
--replace-refs default to 'update-no-add', 2024-07-17), while a step in
the right direction, is prone to leaving around useless replace refs.
Let's update the default to just clean things up, but still leave the
'old-default' option around for folks to get the pre-2024 behavior.

Signed-off-by: Elijah Newren <newren@gmail.com>