filter-repo: provide warnings about filterability on subpar filesystems
If the user has a repo which has refs which differ in case only, then
they will be unable to do filtering on a case insensitive filesystem;
they will get the common
    fatal: cannot lock ref
message when fast-import attempts to lock both refs.  Similarly, if
the user has a repo which has refs which differ in normalization only,
then they will be unable to do filtering on a normalizing filesystem
and will get the same error.

Add a sanity check for both cases when config notifies that we're on
one of the relevant filesystems, and which tells the user which refs
are the problematic ones.

Signed-off-by: Elijah Newren <newren@gmail.com>