filter-repo: explain data structures in AncestryGraph
Each AncestryGraph (both for the original fast export stream, and the
modified stream we write to fast-import) each uses two sets of
identifiers for each commit (giving us a total of four identifiers for a
"commit").  There are good reasons why this is done, but it's easy to
forget and is not well documented.  Add some documentation so that when
I come back again in another five years, I'll be able to quickly figure
out how the class is used again.

Signed-off-by: Elijah Newren <newren@gmail.com>