filter-repo: remove/rename ID_TO_HASH and HASH_TO_ID
ID_TO_HASH and HASH_TO_ID are confusing in that it is difficult to
remember which of the four IDs (two in _orig_graph and two in _graph)
are being referenced.  And trying to use them anywhere new runs into all
kinds of problems trying to figure out what is the correct one to use.

Since these were primarily used with commits, and since the previous
commit started tracking original commit hashes in AncestryGraph, use
that with RepoFilter._orig_graph instead.

For blobs, we do still need to use HASH_TO_ID, but let's rename it to
make it clear which id we are referring to and also scope it to make it
clear that it is only about Blobs.

Signed-off-by: Elijah Newren <newren@gmail.com>