filter-repo: fix gnarly bug in _IDs.record_rename
When a commit is pruned, we don't output it to the fast-import stream.
However, that means the ID of that commit is non-existent, and any
commit which had the pruned commit as a parent needs to instead be
rewritten to have the first unpruned ancestor of the pruned commit as a
parent.  The _IDs class helps with this.

The _IDs class also helps us when random objects are inserted into the
stream, meaning we need to increment the ID of every object in the
stream.  The `lint-history` script in contrib is a good example that
does this.

Now, combining the two causes a corner-case problem because of the
following:
  * When we create a commit, we immediately record a rename if the
    old_id and new id do not match.
  * If we prune a commit, we record a rename so that commits with it
    as a parent instead get its most recent unpruned ancestor

The problem happens as follows:
  * Say we have a stream with commit IDs 1, 2, 3, & 4.
  * The program inserts a blob between each commit, so these commit
    IDs have to be remapped to 1, 3, 5, & 7.
  * We prune the third commit.
In such a case, when we create the third commit, we would immediately
record a remapping of 3->5.  Then, when we later discover after
callbacks that the commit should be pruned, we call record_rename to
update the mapping to instead be 3->3.  The old code would notice our
little optimization that 3==3 and figure that no rename needs to be
recorded.  That would leave the fourth commit trying to use "5" as its
ancestor, which wouldn't exist.

So, while we like avoiding recording renames in general when old_id ==
new_id, if we already have a rename for old_id, then we absolutely need
to update it.

Signed-off-by: Elijah Newren <newren@gmail.com>