git-filter-repo.txt: correct location of metadata files
The explanation of the commit-map and ref-map files from commit
f170a5e61288 (git-filter-repo.txt: document mapping output, 2020-06-18)
included an implicit assumption that "${GIT_DIR}" == ".git/".  While
this may often be true, filtering can also be done in a bare repository,
for example, so this is misleading.  Correct the wording.

Signed-off-by: Elijah Newren <newren@gmail.com>