filter-repo: start tracking new_id -> new_hash mapping as well
As a follow-up to commit dbed13cf375f (filter-repo: allow tracking
commit hashes in AncestryGraph, 2024-09-09), start enabling the tracking
of the mapping between new id and the rewritten hash as well.  To do
so, track more information with _pending_renames so that once the
rename completes we can not only track old_hash -> new_hash, but
new_id -> new_hash as well.

Signed-off-by: Elijah Newren <newren@gmail.com>