filter-repo: rename old_1st_parent -> prev_1st_parent
In _tweak_commit() we had a local variable named old_1st_parent.  This
name was misleading because we often use "old" to refer to objects from
the original fast export stream, and "new" to refer to the new
identifiers used with the new fast import stream (due to insertions of
new blobs or commits, the identifier for a given blob or commit has
different "old" and "new" values).  old_1st_parent, though, was clearly
referring to a "new" identifier for a commit since it was part of
parents rather than orig_parents.  Update the variable name.

Signed-off-by: Elijah Newren <newren@gmail.com>