fixup! filter-repo: add a --file-info-callback
Don't document the delete-the-file; only document dropping the file
change.  As Peff says, it's hard to think of a usecase for it.  But,
by leaving the code in place, if one ever does come up, and someone
can argue for it well, I won't have to cut a new release I can just
point them at the undocumented feature.

And if no one comes up with a usecase for long enough, maybe we can
just drop it.